requirements.txt
fastapi
mysql_connector_repackaged
pydantic
PyJWT
python_bcrypt
uvicorn
